+++ Remote Habits/AppDelegate.swift
@@ -15,12 +15,22 @@ import UserNotifications
 class AppDelegate: UIResponder, UIApplicationDelegate {
 
 
-
+    var userManager = DI.shared.userManager
     func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
         // Override point for customization after application launch.
-
+        
+        updateWorkspaceInfo()
+        guard let workspaceId = userManager.workspaceID, let apiKey = userManager.apiKey else {
+            return true
+        }
         // Step 1: Initialise CIO SDK
-        CustomerIO.initialize(siteId: Env.customerIOSiteId, apiKey: Env.customerIOApiKey, region: Region.US)
+        CustomerIO.initialize(siteId: workspaceId, apiKey: apiKey, region: Region.US)
+               
+        // Step 2: To display rich push notification
+        UNUserNotificationCenter.current().delegate = self
+        
+        // Step 3: Register for push notifications
+        UIApplication.shared.registerForRemoteNotifications()
         return true
     }
 
@@ -38,6 +48,54 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
         // Use this method to release any resources that were specific to the discarded scenes, as they will not return.
     }
 
-
+    // PUSH NOTIFICATIONS
+    func application(_ application: UIApplication, didRegisterForRemoteNotificationsWithDeviceToken deviceToken: Data) {
+        MessagingPush.shared.application(application, didRegisterForRemoteNotificationsWithDeviceToken: deviceToken)
+    }
+     
+    func application(_ application: UIApplication, didFailToRegisterForRemoteNotificationsWithError error: Error) {
+        MessagingPush.shared.application(application, didFailToRegisterForRemoteNotificationsWithError: error)
+    }
+    
+    func updateWorkspaceInfo() {
+        if let workspaceId = userManager.workspaceID {
+            Env.customerIOSiteId = workspaceId
+        }else {
+            userManager.workspaceID = Env.customerIOSiteId
+        }
+        
+        if let apiKey = userManager.apiKey {
+            Env.customerIOApiKey = apiKey
+        }else {
+            userManager.apiKey = Env.customerIOApiKey
+        }
+    }
 }
 
+extension AppDelegate: UNUserNotificationCenterDelegate {
+    func userNotificationCenter(
+        _ center: UNUserNotificationCenter,
+        didReceive response: UNNotificationResponse,
+        withCompletionHandler completionHandler: @escaping () -> Void
+    ) {
+        let handled = MessagingPush.shared.userNotificationCenter(center, didReceive: response,
+                                                                 withCompletionHandler: completionHandler)
+   
+        // If the Customer.io SDK does not handle the push, it's up to you to handle it and call the
+        // completion handler. If the SDK did handle it, it called the completion handler for you.
+        if !handled {
+            completionHandler()
+        }
+    }
+   
+    // OPTIONAL: If you want your push UI to show even with the app in the foreground, override this function and call
+    // the completion handler.
+    @available(iOS 10.0, *)
+    func userNotificationCenter(
+        _ center: UNUserNotificationCenter,
+        willPresent notification: UNNotification,
+        withCompletionHandler completionHandler: @escaping (UNNotificationPresentationOptions) -> Void
+    ) {
+        completionHandler([.list, .banner, .badge, .sound])
+    }
+}