+++ Remote Habits/Model/Profile.swift
@@ -3,3 +3,12 @@ import Foundation
 struct Profile {
     let email: EmailAddress
 }
+
+struct ValidateWorkspaceResponse: Codable {
+    let meta: CredMessage?
+}
+
+struct CredMessage : Codable {
+    let message : String?
+    let error : String?
+}