+++ Remote Habits/View/RHSwitchWorkspaceViewController.swift
@@ -15,7 +15,11 @@ class RHSwitchWorkspaceViewController: RHBaseViewController, UITextFieldDelegate
     @IBOutlet weak var siteIdInput: SkyFloatingLabelTextField!
     @IBOutlet weak var switchWorkspaceButton: UIButton!
     @IBOutlet weak var mainView: UIView!
-    
+    @IBOutlet weak var errorLabel: UILabel!
+
+    // MARK: - VARIABLES
+    var profileViewModel = DI.shared.profileViewModel
+    var userManager = DI.shared.userManager
     // MARK: - --LIFECYCLE METHODS--
     override func viewDidLoad() {
         super.viewDidLoad()
@@ -26,6 +30,7 @@ class RHSwitchWorkspaceViewController: RHBaseViewController, UITextFieldDelegate
         setUpMainView()
         setUpWorkspaceButton()
         setUpTextFields()
+        setUpLabels()
     }
     
     
@@ -35,10 +40,21 @@ class RHSwitchWorkspaceViewController: RHBaseViewController, UITextFieldDelegate
     }
     
     func setUpWorkspaceButton() {
-        switchWorkspaceButton.titleLabel?.font = RHFont.SFProTextSemiBoldMedium
-        switchWorkspaceButton.setTitleColor(UIColor.lightGray, for: .disabled)
-        switchWorkspaceButton.setTitleColor(UIColor.white, for: .normal)
         switchWorkspaceButton.isEnabled = false
+        switchWorkspaceButton.titleLabel?.font = RHFont.SFProTextSemiBoldMedium
+        switchWorkspaceButton.setTitleColor(RHColor.TextDisabled, for: .disabled)
+        switchWorkspaceButton.setTitleColor(UIColor.white, for: .normal)
+        switchWorkspaceButton.layer.cornerRadius = 24
+        workspaceButtonState()
+    }
+    
+    func workspaceButtonState() {
+        
+        switchWorkspaceButton.backgroundColor = switchWorkspaceButton.isEnabled ? RHColor.ButtonEnabled : RHColor.ButtonDisabled
+    }
+    
+    func setUpLabels() {
+        errorLabel.isHidden = true
     }
 
     func setUpTextFields() {
@@ -55,6 +71,8 @@ class RHSwitchWorkspaceViewController: RHBaseViewController, UITextFieldDelegate
     }
     
     func didValueChange(_ sender: Any, otherValue : String) {
+        errorLabel.isHidden = true // In case, error label is visible, hide as soon as user starts typing
+        
         if let floatingLabelTextField = sender as? SkyFloatingLabelTextField {
             if let text = floatingLabelTextField.text {
                 if(!validateInput(with: text)) {
@@ -70,6 +88,8 @@ class RHSwitchWorkspaceViewController: RHBaseViewController, UITextFieldDelegate
                 }else {
                     switchWorkspaceButton.isEnabled = false
                 }
+                
+                workspaceButtonState()
             }
         }
     }
@@ -87,7 +107,21 @@ class RHSwitchWorkspaceViewController: RHBaseViewController, UITextFieldDelegate
         // Pass the selected object to the new view controller.
     }
     */
+    func routeToLogin() {
+        if let navController = self.parent as? UINavigationController, let presenter = navController.presentingViewController as? UINavigationController {
+            
+            let main = UIStoryboard(name: RHConstants.kStoryboardMain, bundle: nil)
+            let vc = main.instantiateViewController(withIdentifier: RHConstants.kLoginViewController) as! RHLoginViewController
+            presenter.setViewControllers([vc], animated: true)
+            self.dismiss(animated: true, completion: nil)
+        }
+    }
     
+    func textFieldShouldReturn(_ textField: UITextField) -> Bool {
+        self.view.endEditing(true)
+        return false
+    }
+
     // MARK: - --ACTIONS--
     
     @IBAction func siteIdValueChanged(_ sender: Any) {
@@ -96,5 +130,31 @@ class RHSwitchWorkspaceViewController: RHBaseViewController, UITextFieldDelegate
     
     @IBAction func apiKeyValueChanged(_ sender: Any) {
         didValueChange(sender, otherValue: siteIdInput.text ?? RHConstants.kEmptyValue)
+    }
+    
+    @IBAction func switchWorkspaceButtonTapped(_ sender: UIButton) {
+        view.endEditing(true) // Close keyboard if it is open
+        showLoadingView()
+        guard let siteId = siteIdInput.text , let apikey = apiKeyInput.text else { return }
+        profileViewModel.validateWorkspace(forSiteId: siteId, and: apikey) { [weak self] result in
+            guard let self = self else { return }
+            
+            self.hideLoadingView()
+            if result {
+                self.profileViewModel.logoutUser()
+                self.updateWorkspaceInfo(with : siteId , andApiKey : apikey)
+                self.routeToLogin()
+            }
+            else {
+                self.errorLabel.isHidden = false
+            }
+        }
+    }
+    
+    private func updateWorkspaceInfo(with siteId : String, andApiKey apiKey : String) {
+        userManager.workspaceID = siteId
+        Env.customerIOSiteId = siteId
+        userManager.apiKey = apiKey
+        Env.customerIOApiKey = apiKey
     }
 }