+++ Remote Habits/View/Xib/HabitDashboard/HabitTableViewCell.swift
@@ -1,5 +1,5 @@
 //
-//  HabitTableViewCell.swift
+//  HabitTableViewswift
 //  Remote Habits Mobile App
 //
 //  Created by Amandeep Kaur on 30/11/21.
@@ -16,6 +16,7 @@ class HabitTableViewCell: UITableViewCell {
     @IBOutlet weak var habitIcon: UIImageView!
     @IBOutlet weak var mainCellView: UIView!
     var actionType : HabitActionType?
+    var habitData : HabitData?
     
     var actionDelegate: RHDashboardActionHandler?
     override func awakeFromNib() {
@@ -30,7 +31,38 @@ class HabitTableViewCell: UITableViewCell {
         // Configure the view for the selected state
     }
     
+    func fillData() {
+        guard let habitData = habitData else {
+            return
+        }
+
+        habitIcon.image = UIImage(named: habitData.icon ?? RHConstants.kLogo)
+        habitTitle.text = habitData.title
+        habitSubTitle.text = habitData.subTitle
+        actionType = habitData.habitDetail?.actionType
+        if let type = habitData.type {
+            if type == .toggleSwitch {
+                habitSwitch.isHidden = false
+                actionButton.isHidden = true
+                habitSwitch.setOn(habitData.habitDetail?.isHabitEnabled ??  false, animated: true)
+            }
+            else if type == .button {
+                actionButton.isHidden = false
+                habitSwitch.isHidden = true
+                actionButton.isEnabled = habitData.habitDetail?.isHabitEnabled ?? true
+                actionButton.setTitle(habitData.habitDetail?.actionButtonValue, for: .normal)
+            }
+        }
+        else {
+            actionButton.setTitle(RHConstants.kEmptyValue, for: .normal)
+            habitSwitch.isHidden = true
+            actionButton.isHidden = true
+        }
+    }
+    
     @IBAction func habitSwitchValueChanged(_ sender: UISwitch) {
+        let selectedHabitData = SelectedHabitData(title: habitData?.title, frequency: habitData?.habitDetail?.frequency, startTime: habitData?.habitDetail?.startTime, endTime: habitData?.habitDetail?.endTime)
+        actionDelegate?.toggleHabit(toValue: sender.isOn, habitData: selectedHabitData)
     }
     
     @IBAction func habitActionButtonTapped(_ sender: UIButton) {